pyproject.toml
[project]
name = "max-resending"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "maxapi-python>=1.1.13",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
